retro_frontend: Remove leaky `Frontend` abstraction
/shrug. Turns out this was pretty easy with how early this was, so I'm glad I did it now *before* I had to deal with the ramifications.

This follows the first path in the previous commit, by removing the Frontend struct entirely. In its place are the same once_cell/unsafe wrappers, but now they don't require self bindings. The implication is that, yes, the frontend is global state, so you can only have *the* frontend in your application.