retro_frontend: Implement ROM loading
Right now, this uses the old shoddy game info structure and method of loading ROMs, as it's the easiest and most widely used.

Also implements more things in the environment callback. We can probably serde settings/variables as TOML or JSON or something like that, but for now, I'm not going to worry about that yet; we need the notion of given directory paths first before worrying about saving data anyways. (also, I imagine allowing another party, i.e: the consumer, to load setting  would be more uniform to the goal of making this a library; we could probably use a marshalled and sanified SET_VARIABLES in some admin panel payload to limit values and such.)