retro_frontend: A bunch of stuff
More callbacks, moved to new module because there are so many. A lot of these will need ot be implemented fully by a consumer, but for now all except video are stubs pretty much.

Also, retrodemo kind of does something now, it runs the core and (tries. key word) to display the result. Doesn't work great, but it's at least doing something.