chore(*): update retro_frontend from retrovnc
OpenGL support. Yay! EGL bindings are in a new crate, and are used by retrodemo to create a headless OpenGL context. (Not that that matters for retrodemo, since it, well, displays a window, but hey, it's there.)

Unfortunately retrodemo is a bit more complex now.. :( I'll have to see if I can clean it up further/split it into modules like I did retrovnc, and split the EGL initalization/finialization into another helper. Or something, I'm spitballing.