letsplay_retro_frontend: Rewrite C++ log helper to use a string pool for formatted log messages
This allows a formatted string to be any length above without any issues. Also avoids buffer overflows (especially of the stack kind.) since we size based on what vsnprintf() returns.

Note that I haven't found a core that would break the previous implementation, but it was a hack and this feels a bit less not like a hack.