letsplay_retro_frontend: Make stringpool do actual garbage collection
Now, if the freelist is considered too large, it will not blindly destroy strings which are marked as in use, but rather only free strings which are not in use. To support this (easily), the freelist is now implemented as a doubly linked list, in which removal of list nodes at any arbitrary point is a LOT less painful.